requirements.txt
discord.py==1.6.0
aiohttp==3.7.3
python-dotenv==0.15.0